Rename lib to cortex
